chore: rebase metrics refactoring changes on main
