--- docs/development/contributing.mdx
@@ -1,195 +1,0 @@
----
-title: "Contributing"
-description: "Guidelines for contributing to OpenBench"
----
-
-Thank you for your interest in contributing to OpenBench!
-This guide covers everything you need to know to contribute effectively.
-
-### Key Development Tools
-
-**uv**: Package manager - use uv add "package>=version" for dependencies  
-**ruff**: Linting and formatting  
-**mypy**: Type checking - required for all new code  
-**pre-commit**: Automated code quality checks - must pass before commits
-**pytest**: Testing framework with integration test markers
-
-## Getting Started
-
-### 1. Fork and Clone
-
-```bash
-# Fork on GitHub, then clone
-git clone https://github.com/YOUR_USERNAME/openbench.git
-cd openbench
-
-# Add upstream remote
-git remote add upstream https://github.com/lelouvincx/openbench.git
-```
-
-### 2. Set Up Development Environment
-
-```bash
-# Create virtual environment with UV
-uv venv
-source .venv/bin/activate  # On Windows: .venv\Scripts\activate
-
-# Install in development mode
-uv sync --dev
-
-# CRITICAL: Install pre-commit hooks
-pre-commit install
-
-# Run tests to verify setup
-pytest
-```
-
-### 3. Create Feature Branch
-
-```bash
-# Update main branch
-git checkout main
-git pull upstream main
-
-# Create feature branch
-git checkout -b feature/your-feature-name
-```
-
-## Make Your Changes
-
-### Core Principles
-
-<AccordionGroup>
-  <Accordion title="Single Responsibility">
-    - Focus on one feature, bug fix, or improvement per pull request
-    - Makes code review easier and reduces merge conflicts
-  </Accordion>
-  
-  <Accordion title="Clear Separation of Concerns (SoC)">
-    - Keep functions small and focused
-    - Separate business logic from presentation logic
-    - Use clear interfaces between components
-    - Avoid tight coupling between modules
-  </Accordion>
-  
-  <Accordion title="Documentation and Readability">
-    - Use type hints for clarity and validation
-    - Write docstrings for all public functions and classes
-    - Use descriptive variable and function names
-  </Accordion>
-  
-  <Accordion title="Standardized Structure">
-    - Follow PEP 8 with a line length of 88 characters
-    - Mirror OpenBench/InspectAI conventions
-    - Use consistent formatting and naming patterns
-    - Maintain consistency with existing codebase
-  </Accordion>
-</AccordionGroup>
-
-### Testing
-
-Write unit tests for all new functionality. Aim for high test coverage including edge cases and error conditions.
-
-```python   tests/test_my_feature.py
-import pytest
-from openbench.my_module import my_function
-
-def test_my_function():
-    """Test my_function with valid input"""
-    result = my_function("input")
-    assert result == "expected"
-
-def test_my_function_edge_case():
-    """Test my_function with edge case"""
-    with pytest.raises(ValueError):
-        my_function(None)
-```
-
-```bash
-# Run all tests
-pytest
-
-# Run specific test file
-pytest tests/test_my_feature.py
-```
-
-### Code Quality Checks
-
-```bash
-# Run all pre-commit hooks
-pre-commit run --all-files
-
-# Individual checks
-ruff check .          # Linting
-ruff format .         # Formatting
-mypy .               # Type checking
-```
-
-## Submitting Changes
-
-### Commits
-
-Follow [conventional commits](https://www.conventionalcommits.org/en/v1.0.0/) structure:  
-`<type>(<scope>): <subject>`
-
-```
-# Examples
-feat(mmlu): add multilingual support
-fix(scorer): handle empty responses correctly
-docs(readme): update installation instructions
-test(humaneval): add integration tests
-refactor(registry): simplify task loading
-chore(deps): update inspect-ai to 0.3.0
-```
-
-### Pull Requests
-
-#### 1. Before Opening PR
-
-```bash
-# Ensure tests pass
-pytest
-
-# Run pre-commit hooks
-pre-commit run --all-files
-
-# Update from upstream
-git fetch upstream
-git rebase upstream/main
-```
-
-#### 2. PR Description Template
-
-```markdown
-## Description
-Brief description of changes
-
-## Motivation
-Why is this change needed?
-
-## Changes
-- Change 1
-- Change 2
-
-## Testing
-How has this been tested?
-
-## Checklist
-- [ ] Tests pass
-- [ ] Pre-commit hooks pass
-- [ ] Documentation updated
-- [ ] Changelog updated (if applicable)
-
-Fixes #issue_number
-```
-
-#### 3. Review Process
-
-1. Maintainers will review your PR
-2. Address feedback promptly
-3. Once approved, we'll squash and merge your PR
-4. Your contribution will be part of the next release!
-
-<br/>
-Thank you for contributing to OpenBench!
-We are grateful for your support in making language model evaluation more accessible and reliable. 🫶
