--- docs/docs.json
@@ -1,124 +1,0 @@
-{
-  "$schema": "https://mintlify.com/docs.json",
-  "theme": "aspen",
-  "name": "OpenBench",
-"colors": {
-  "primary": "#C17559",
-  "light": "#4A90E2",
-  "dark": "#2D3748"
-},
-  "fonts": {
-    "family": "Neue Regrade",
-    "weight": 500,
-    "heading": {
-      "family": "Neue Regrade",
-      "weight": 700
-    },
-    "body": {
-      "family": "Neue Regrade",
-      "weight": 500
-    }
-  },
-  "favicon": "/favicon.svg",
-  "navigation": {
-    "groups": [
-      {
-        "group": "Getting Started",
-        "pages": [
-          "index",
-          "installation",
-          "quickstart",
-          "configuration",
-          "benchmarks_catalog",
-          "troubleshooting"
-        ]
-      },
-      {
-        "group": "CLI Reference",
-        "pages": [
-          "cli/overview",
-          "cli/eval",
-          "cli/eval-retry",
-          "cli/list",
-          "cli/describe",
-          "cli/view"
-        ]
-      },
-      {
-        "group": "Development",
-        "pages": [
-          "development/architecture",
-          "development/contributing"
-        ]
-      }
-    ],
-    "global": {
-      "anchors": [
-        {
-          "anchor": "GitHub",
-          "href": "https://github.com/groq/openbench",
-          "icon": "github"
-        },
-        {
-          "anchor": "PyPI",
-          "href": "https://pypi.org/project/openbench/",
-          "icon": "python"
-        },
-        {
-          "anchor": "Groq",
-          "href": "https://groq.com",
-          "icon": "microchip"
-        }
-      ]
-    }
-  },
-  "logo": {
-    "light": "/logo/light.svg",
-    "dark": "/logo/dark.svg"
-  },
-  "navbar": {
-    "links": [
-      {
-        "label": "Issues",
-        "href": "https://github.com/groq/openbench/issues"
-      }
-    ],
-    "primary": {
-      "type": "button",
-      "label": "Get Started",
-      "href": "https://github.com/groq/openbench"
-    }
-  },
-  "contextual": {
-    "options": [
-     "copy",
-     "view",
-     "chatgpt",
-     "claude",
-     "perplexity",
-     "mcp",
-     "cursor",
-     "vscode"
-   ]
-  },
-  "footer": {
-    "socials": {
-      "x": "https://x.com/GroqInc",
-      "github": "https://github.com/groq/openbench",
-      "linkedin": "https://linkedin.com/company/groq"
-    }
-  },
-  "analytics": {
-    "plausible": {
-      "domain": "openbench.groq.com"
-    }
-  },
-  "modeToggle": {
-    "default": "dark"
-  },
-  "feedback": {
-    "thumbsRating": true,
-    "suggestEdit": true,
-    "raiseIssue": true
-  }
-}
