--- docs/fonts/README.md
@@ -1,12 +1,0 @@
-# Neue Regrade font
-
-This folder is where the Neue Regrade font file should live for your docs.
-
-Expected filename (used by style.css and docs.json):
-
-- `NeueRegrade-Variable.ttf`
-
-Drop the variable TTF here. The site will load the font automatically via
-`style.css` and the `fonts` section in `docs.json`.
-
-If the file is missing, the site will gracefully fall back to system fonts.
