--- docs/style.css
@@ -1,13 +1,0 @@
-/* Neue Regrade font-face (variable). Place the TTF at /fonts/NeueRegrade-Variable.ttf */
-@font-face {
-  font-family: "Neue Regrade";
-  src: url("/fonts/NeueRegrade-Variable.ttf") format("truetype");
-  font-weight: 300 800; /* variable weight range */
-  font-style: normal;
-  font-display: swap;
-}
-
-/* Optional: ensure the body picks up the configured font if available */
-body {
-  font-family: "Neue Regrade", system-ui, -apple-system, Segoe UI, Roboto, Helvetica, Arial, sans-serif;
-}
